% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_uncertainty.R
\name{cum_uncertainty}
\alias{cum_uncertainty}
\title{Cumulative uncertainty(DEPRECATED)}
\usage{
cum_uncertainty(data, var_name, stages = setdiff(names(data), var_name),
  U = var0)
}
\arguments{
\item{data}{a data frame containing scenarios(factor or character) for each stages and the variable of interest(numeric).
data should contain all combinations of scenarios.}

\item{var_name}{the name of the variable of interest}

\item{stages}{names of the stages in the modeling chain. should be ordered by the order of the modeling chain}

\item{U}{a function that returns uncertainty such as range and variance of a given numeric vector.
This package have built-in uncertainty functions var0() and drange(). Default is var0().}
}
\value{
summary of uncertainties
}
\description{
This function performs uncertainty decomposition based on the cumulative uncertainty.
}
\examples{
stage1 <- LETTERS[1:3]
stage2 <- LETTERS[1:2]
stage3 <- LETTERS[1:4]
y <- rnorm(3*2*4)
data <- expand.grid(stage1=stage1,
                    stage2=stage2,
                    stage3=stage3)
data <- cbind(data, y)
# cum_uncertainty() is deprecated. Use UD_cum_stage()
# cum_uncertainty(data,"y", names(data)[-4])
# cum_uncertainty(data,"y", names(data)[-4],drange)
}
