% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drange.R
\name{drange}
\alias{drange}
\title{Range}
\usage{
drange(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
the difference of maximum and minimum of a given vector
}
\description{
This function returns the difference of maximum and minimum of a given vector.
}
\examples{
(x <- rnorm(5))
#drange(x)
}
