% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unc_measures.R
\docType{data}
\name{unc_measures}
\alias{unc_measures}
\alias{U_range}
\alias{u_range}
\alias{flist_range}
\alias{U_mad}
\alias{u_mad}
\alias{flist_mad}
\alias{U_var}
\alias{u_var}
\alias{flist_var}
\title{uncertainty measures}
\format{An object of class \code{list} of length 3.}
\usage{
U_range(x)

u_range(x, a, b, n)

flist_range

U_mad(x)

u_mad(x, m)

flist_mad

U_var(x)

u_var(x, m)

flist_var
}
\arguments{
\item{x}{a vector.}

\item{a, b, n}{summary statistics for u_range(), a and b are minimum and maximum of x. n is length of x}

\item{m}{summary statistics. median for u_mad() and mean for u_var()}
}
\description{
Functions beginning with U are uncertainty measure that return a scalar given a vector, such as mean absolute deviation or variance. Functions beginning with flist are lists of functions that summarize vector like mean or median. 
Functions beginning with u are the uncertainty of each element of the vector, which is calculated from the elements of the vector and summary statistics of the vector, like difference or square of difference between two values.
}
\keyword{datasets}
