% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var0.R
\name{var0}
\alias{var0}
\title{Variance}
\usage{
var0(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
the population variance of a given vector
}
\description{
This function returns the population variance of a given vector.
}
\examples{
(x <- rnorm(5))
# var0(x)
}
